+++ frontend/src/Components/Login.js
@@ -15,24 +15,31 @@ function Login() {
     let navigate = useNavigate()
     const loginUrl = "http://localhost:8000/user/"
 
+
+
+
     useEffect(() => {
         console.log("rendering useEffect")
-        console.log(logSubmissionStatus)
-        alert("congratuatlations")
-
         const payload = { "email": email, "password": pwd }
         if (email) {
-            console.log("Sending Axios request")
             axios.post(loginUrl, payload).then((resp) => {
-                alert("congratuatlations from within")
                 setLoginSubmissionStatus("completed")
                 setLoginMessage(`You have succesfully created a user account with the credentials ${email}`);
                 navigate("/user_profile");
 
             }).catch(err => {
-                if (logSubmissionStatus == "ready") {
-                    setLoginMessage(`${JSON.parse(err.request.response).detail}! `)
+                console.log(err)
+                if (err.code == "ERR_NETWORK"){
+                    setLoginMessage(`Server is down. Please try again later.`)
+                    return
                 }
+                if (logSubmissionStatus == "ready") {
+                    setLoginMessage(`${JSON.parse(err.request.response).detail}`)
+                    setEmail("")
+                    setPwd("")
+                    setLoginSubmissionStatus("pending")
+                }
+                    
             }
 
             )
@@ -40,10 +47,12 @@ function Login() {
 
     }, [logSubmissionStatus])
 
+
     // enter data in form
     const readCredentials = () => {
 
         enteredEmail = document.querySelector("#username_login").value
+        console.log("enteredEmail", enteredEmail)
         enteredPwd = document.querySelector("#pwd_login").value
         let emailRegex = /^[a-zA-Z0-9]+@[a-zA-Z]+\.[a-z]+$/
         let pwdRegex = /^[a-zA-Z0-9!]{3,12}$/
@@ -71,7 +80,7 @@ function Login() {
                 <form className={formStyle}>
                     <input className={formElementStyle} id="username_login" type="text" name="username" autoComplete="off" onClick={() => setLoginSubmissionStatus("pending")} placeholder="Enter your email" pattern="^[a-zA-Z0-9]+@[a-zA-Z]+\.[a-z]+$" title="Please enter a valid email address" required />
                     <input className={formElementStyle} id="pwd_login" type="password" name="pwd" placeholder="Enter your password" onClick={() => setLoginSubmissionStatus("pending")} pattern="^[a-zA-Z0-9!]{3,12}$" title="Please enter a password with length between 3 and 12 characters" required />
-                    <button className={buttonClass} name="login" type="submit" onClick={readCredentials}>Login</button>
+                    <button className={buttonClass} name="login" type="button" onClick={readCredentials}>Login</button>
                     <button className={buttonClass} name="signup" type="button" onClick={() => navigate("/signup")}>Sign up</button>
 
                 </form>